TVIST1-579: Reduce code duplication and increase readability
