+++ src/Controller/Admin/MailTemplateCrudController.php
@@ -57,8 +57,12 @@ class MailTemplateCrudController extends AbstractCrudController
         ;
         yield TextField::new('name');
         yield TextareaField::new('customFields', 'Custom fields')
-            ->setHelp($this->translator->trans('List of custom fields(one per line). Should not contain any spaces.',
+            ->setHelp($this->translator->trans('List of custom fields (one per line). Should not contain any spaces.',
                 [], 'admin'))
+            // Avoid showing 'Null' or 'Tom' if customFields is not set.
+            ->formatValue(function ($value) {
+                return $value ?? ' ';
+            })
         ;
         yield TextareaField::new('description');
         yield Field::new('templateFile')
