+++ src/Controller/AgendaBroadcastController.php
@@ -60,7 +60,6 @@ class AgendaBroadcastController extends AbstractController
 
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
-            // Handle potential merge fields
             $customData = [];
 
             foreach ($form->get('customData') as $customField) {
