+++ src/Controller/AgendaInspectionController.php
@@ -65,7 +65,6 @@ class AgendaInspectionController extends AbstractController
 
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
-            // Handle potential merge fields
             $customData = [];
 
             foreach ($form->get('customData') as $customField) {
