+++ src/Entity/AgendaBroadcast.php
@@ -3,9 +3,9 @@
 namespace App\Entity;
 
 use App\Repository\AgendaBroadcastRepository;
+use App\Traits\CustomDataTrait;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Timestampable\Traits\TimestampableEntity;
-use Symfony\Component\Serializer\Annotation\Groups;
 use Symfony\Component\Uid\Uuid;
 
 /**
@@ -13,6 +13,7 @@ use Symfony\Component\Uid\Uuid;
  */
 class AgendaBroadcast
 {
+    use CustomDataTrait;
     use TimestampableEntity;
 
     /**
@@ -42,12 +43,6 @@ class AgendaBroadcast
      * @ORM\JoinColumn(nullable=false)
      */
     private $agenda;
-
-    /**
-     * @ORM\Column(type="json")
-     * @Groups({"mail_template"})
-     */
-    private $customData = [];
 
     public function __toString()
     {
@@ -108,18 +103,6 @@ class AgendaBroadcast
     public function setAgenda(?Agenda $agenda): self
     {
         $this->agenda = $agenda;
-
-        return $this;
-    }
-
-    public function getCustomData(): ?array
-    {
-        return $this->customData;
-    }
-
-    public function setCustomData(?array $customData): self
-    {
-        $this->customData = $customData;
 
         return $this;
     }
