+++ src/Form/HearingPostRequestType.php
@@ -3,21 +3,19 @@
 namespace App\Form;
 
 use App\Entity\HearingPostRequest;
-use App\Entity\MailTemplate;
-use Symfony\Bridge\Doctrine\Form\Type\EntityType;
+use App\Traits\TemplateFormTrait;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\CollectionType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\Form\FormEvent;
-use Symfony\Component\Form\FormEvents;
-use Symfony\Component\Form\FormInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 class HearingPostRequestType extends AbstractType
 {
+    use TemplateFormTrait;
+
     public function __construct(private TranslatorInterface $translator)
     {
     }
@@ -37,50 +35,14 @@ class HearingPostRequestType extends AbstractType
         $caseParties = $options['case_parties'];
         $availableTemplateChoices = $options['mail_template_choices'];
 
-        $templateChoices = [];
-
-        foreach ($availableTemplateChoices as $template) {
-            $templateChoices[$template->getName()] = $template;
-        }
-
         $builder
             ->add('title', TextType::class, [
                 'label' => $this->translator->trans('Title', [], 'case'),
                 'help' => $this->translator->trans('Choose a title for the hearing post', [], 'case'),
             ])
-            ->add('template', EntityType::class, [
-                'class' => MailTemplate::class,
-                'placeholder' => $this->translator->trans('Choose a template', [], 'case'),
-                'label' => $this->translator->trans('Mail template', [], 'case'),
-                'choices' => $templateChoices,
-            ])
         ;
 
-        $formModifier = function (FormInterface $form, MailTemplate $mailTemplate = null) use ($builder) {
-            $form->add('customData', MailTemplateCustomDataType::class, [
-                'label' => false,
-                'template' => $mailTemplate,
-                'data' => $builder->getData()->getCustomData(),
-            ]);
-        };
-
-        $builder->addEventListener(
-            FormEvents::PRE_SET_DATA,
-            function (FormEvent $event) use ($formModifier) {
-                $data = $event->getData();
-
-                $formModifier($event->getForm(), $data->getTemplate());
-            }
-        );
-
-        $builder->get('template')->addEventListener(
-            FormEvents::POST_SUBMIT,
-            function (FormEvent $event) use ($formModifier) {
-                $template = $event->getForm()->getData();
-
-                $formModifier($event->getForm()->getParent(), $template);
-            }
-        );
+        $this->addTemplate($builder, $availableTemplateChoices);
 
         $builder
             ->add('recipient', ChoiceType::class, [
