+++ translations/admin.da.xlf
@@ -345,13 +345,13 @@
         <source>Select receipt template</source>
         <target state="translated">Vælg kvitteringsskabelon</target>
       </trans-unit>
-      <trans-unit id="MtlQbTd" resname="List of custom fields(one per line). Should not contain any spaces.">
-        <source>List of custom fields(one per line). Should not contain any spaces.</source>
-        <target state="translated">Liste af brugerdefinerede flettefelter (en pr. linje). Må ikke indeholde mellemrum.</target>
-      </trans-unit>
       <trans-unit id="VWbbSfr" resname="Custom fields">
         <source>Custom fields</source>
         <target state="translated">Brugerdefinerede flettefelter</target>
+      </trans-unit>
+      <trans-unit id="k9uhES_" resname="List of custom fields (one per line). Should not contain any spaces.">
+        <source>List of custom fields (one per line). Should not contain any spaces.</source>
+        <target state="translated">Liste af brugerdefinerede flettefelter (en pr. linje). Må ikke indeholde mellemrum.</target>
       </trans-unit>
     </body>
   </file>
